Add `BookEv` firing - major testing problems
The way I thought we could test the connector seems to have hit a major
roadblock. It seems I didn't think about it clearly enough before.

The `TimedLogger p v` type is the "state part" of a State Monad that we were going
to use for testing (to avoid using IO).

However, I wanted to make the p and v parameters in the `TimedLogger p v` type match the same
parameters in the Exchange class signature

class Exchange config m where
    placeLimit    :: (HTTP m, MonadTime m, Coin p, Coin v) => config -> OrderSide -> Price p -> Vol v -> m OrderID
    ...

But those p and v are different.

So, we have an extra "forall". A given `TimedLogger p v` fixes a given p and v and thus, this type canNOT have a
`Exchange Coinbene` instance because in an exchange instance a single monadic type works for all markets. The
`TimedLogger p v` would only work for a single market.

There is also a problem with IO. We will need to interweave IO actions within a stateful computation. The current
IntoIO mechanism seems insufficient for this.

I have to think of a different way to make the testing work.