WIP: Enable parametricity in tests
We can now write simple test cases that will work for any Coin type.
This is required to allow us to write a mock version of the `Exchange` interface
with specific responses.

This passes the test by cheating. I now have to change to code and update the
test cases to keep state so that we can provide different responses through the
`Exchange` API.