WIP: Implement "all but the brains" of the producer
The logic for event detection in the producer is missing only its
"inner most loop" (`updateOrder`) that compares new order information
and the orders current recorded state and decides if/how to update
the state and what events to dispatch. All else is implemented.

Lots of cleaning up and testing still required.